Visual Studio Project with test and results
Unfortunately back when the code for this project was written I was not particularly familiar with github and hence did not use it, therefore this commit consists of the entire finished project rather than each step taken towards its completion. However, this project is for the most part based on the MINIC compiler(with the necessary changes made in the Parser, the Lexer and so forth), so through studying the steps that are detailed on the MINIC project and comparing its files with the Mini-Python ones, you should be able to understand the procedure.